fix: replace Vercel deployment webhook with GitHub push-event webhook and remove insecure default secret
Addresses reviewer feedback:
1. Changed endpoint from /api/webhooks/deployments to /api/webhooks/github
2. Handles GitHub push events instead of Vercel deployments (per issue #1935)
3. Removed insecure hardcoded fallback secret - returns 503 if env var missing
4. Uses crypto.timingSafeEqual for constant-time signature comparison
5. Updated lockfile with new dependencies
