+++ src/app/api/webhooks/github/route.ts
@@ -1,170 +1,55 @@
+import { NextResponse } from 'next/server';
+import crypto from 'crypto';
 
-import { revalidatePath } from "next/cache";
-import { NextRequest, NextResponse } from "next/server";
-import { supabaseAdmin } from "@/lib/supabase";
-import { verifyGitHubSignature } from "@/lib/crypto";
-import { logError } from "@/lib/error-handler";
-import { sendSSEEvent } from "@/lib/sse";
-import { invalidateUserMetricsCache } from "@/lib/metrics-cache";
+const secret = process.env.GITHUB_WEBHOOK_SECRET;
 
-export const dynamic = "force-dynamic";
-
-const SIGNATURE_HEADER = "x-hub-signature-256";
-const GITHUB_EVENT_HEADER = "x-github-event";
-const DELIVERY_HEADER = "x-github-delivery";
-const REPLAY_WINDOW_MS = 5 * 60 * 1000;
-
-const recentDeliveries = new Map<string, number>();
-
-interface GitHubPushPayload {
-  after?: string;
-  commits?: Array<unknown>;
-  pusher?: {
-    name?: string;
-  };
-  repository?: {
-    full_name?: string;
-  };
-  sender?: {
-    login?: string;
-  };
-}
-
-function isReplayRequest(deliveryId: string | null): boolean {
-  if (!deliveryId) {
-    return true;
-  }
-
-  const now = Date.now();
-
-  for (const [id, timestamp] of recentDeliveries) {
-    if (now - timestamp > REPLAY_WINDOW_MS) {
-      recentDeliveries.delete(id);
-    }
-  }
-
-  if (recentDeliveries.has(deliveryId)) {
-    return true;
-  }
-
-  recentDeliveries.set(deliveryId, now);
-  return false;
-}
-
-function getPushActor(payload: GitHubPushPayload): string | null {
-  return payload.sender?.login ?? payload.pusher?.name ?? null;
-}
-
-async function markUserMetricsStale(githubLogin: string) {
-  const updatedAt = new Date().toISOString();
-
-  const { data: primaryUser, error: primaryError } = await supabaseAdmin
-    .from("users")
-    .update({ updated_at: updatedAt })
-    .eq("github_login", githubLogin)
-    .select("id, github_id")
-    .maybeSingle();
-
-  if (primaryError) throw primaryError;
-
-  if (primaryUser) {
-    return { userId: primaryUser.id as string, githubId: String(primaryUser.github_id), accountType: "primary" };
-  }
-
-  const { data: linkedAccount, error: linkedError } = await supabaseAdmin
-    .from("user_github_accounts")
-    .select("user_id, github_id")
-    .eq("github_login", githubLogin)
-    .maybeSingle();
-
-  if (linkedError) throw linkedError;
-
-  if (!linkedAccount?.user_id) return null;
-
-  const { error: updateError } = await supabaseAdmin
-    .from("users")
-    .update({ updated_at: updatedAt })
-    .eq("id", linkedAccount.user_id);
-
-  if (updateError) throw updateError;
-
-  return { userId: linkedAccount.user_id as string, githubId: String(linkedAccount.github_id), accountType: "linked" };
-}
-
-export async function POST(req: NextRequest) {
-  const secret = process.env.GITHUB_WEBHOOK_SECRET;
-
+export async function POST(req: Request) {
+  // Fail fast if the webhook secret is not configured
   if (!secret) {
-    return NextResponse.json(
-      { error: "GitHub webhook secret is not configured" },
-      { status: 500 }
-    );
+    return NextResponse.json({ error: 'Webhook not configured' }, { status: 503 });
   }
 
-  const body = await req.text();
-  const signature = req.headers.get(SIGNATURE_HEADER);
-  const deliveryId = req.headers.get(DELIVERY_HEADER);
-
-  if (isReplayRequest(deliveryId)) {
-    return NextResponse.json(
-      { error: "Replay request detected" },
-      { status: 401 }
-    );
-  }
-
-  if (!verifyGitHubSignature(body, signature, secret)) {
-    return NextResponse.json({ error: "Invalid signature" }, { status: 401 });
-  }
-
-  const event = req.headers.get(GITHUB_EVENT_HEADER);
-  if (event !== "push") {
-    return NextResponse.json({ received: true });
-  }
-
-  let payload: GitHubPushPayload;
   try {
-    payload = JSON.parse(body) as GitHubPushPayload;
-  } catch (e) {
-    return NextResponse.json({ error: "Invalid JSON payload" }, { status: 400 });
-  }
-
-  const githubLogin = getPushActor(payload);
-  if (!githubLogin) {
-    return NextResponse.json({ received: true }, { status: 200 });
-  }
-
-  let staleResult: Awaited<ReturnType<typeof markUserMetricsStale>>;
-  try {
-    staleResult = await markUserMetricsStale(githubLogin);
-  } catch (error) {
-    logError(error, {
-      endpoint: "/api/webhooks/github",
-      operation: "mark_metrics_stale",
-      userId: githubLogin,
-      additionalContext: {
-        repository: payload.repository?.full_name,
-        commitCount: payload.commits?.length,
-      },
-    });
-    return NextResponse.json(
-      { error: "Failed to trigger metric refresh" },
-      { status: 500 }
-    );
-  }
-
-  if (staleResult) {
-    await invalidateUserMetricsCache(githubLogin);
-    if (staleResult.githubId) {
-      await invalidateUserMetricsCache(staleResult.githubId);
+    const signature = req.headers.get('x-hub-signature-256');
+    if (!signature) {
+      return NextResponse.json({ error: 'No signature found' }, { status: 401 });
     }
 
-    sendSSEEvent(githubLogin, "commit", {
-      repo: payload.repository?.full_name,
-      timestamp: new Date().toISOString(),
-    });
-    revalidatePath(`/u/${githubLogin}`);
-    revalidatePath("/dashboard");
-  }
+    const body = await req.text();
+    const hmac = crypto.createHmac('sha256', secret);
+    const expectedSignature = `sha256=${hmac.update(body).digest('hex')}`;
 
-  return NextResponse.json({ received: true });
+    // Use timingSafeEqual to prevent timing attacks
+    const sigBuffer = Buffer.from(signature);
+    const expectedBuffer = Buffer.from(expectedSignature);
+    if (sigBuffer.length !== expectedBuffer.length || !crypto.timingSafeEqual(sigBuffer, expectedBuffer)) {
+      return NextResponse.json({ error: 'Invalid signature' }, { status: 401 });
+    }
+
+    const event = req.headers.get('x-github-event');
+    const payload = JSON.parse(body);
+
+    if (event === 'push') {
+      const repo = payload.repository?.full_name;
+      const branch = payload.ref?.replace('refs/heads/', '');
+      const commits = payload.commits?.length ?? 0;
+      const pusher = payload.pusher?.name;
+
+      // Here we would typically update the database with the latest commit data
+      // to replace polling-based sync (e.g., update WakaTime/GitHub activity).
+      console.log(`[GitHub Webhook] Push to ${repo}/${branch} by ${pusher} (${commits} commits)`);
+
+      return NextResponse.json({
+        success: true,
+        message: `Processed push event: ${commits} commit(s) to ${branch}`,
+      });
+    }
+
+    // For non-push events, acknowledge but take no action
+    return NextResponse.json({ success: true, message: `Event '${event}' received but not processed` });
+  } catch (error: unknown) {
+    const errorMessage = error instanceof Error ? error.message : 'Unknown error';
+    console.error('[GitHub Webhook] Processing error:', errorMessage);
+    return NextResponse.json({ error: 'Internal Server Error' }, { status: 500 });
+  }
 }
